project/requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.3

# Database connectors
duckdb==0.9.2
google-cloud-bigquery==3.14.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
pyarrow==14.0.2

# Visualization
plotly==5.18.0

# Environment management
python-dotenv==1.0.0
